cleaning up some syntax issues

+++ TestSuite/Source/GetSchemaTests.cs
@@ -46,7 +46,7 @@ namespace MySql.Data.MySqlClient.Tests
 			Assert.AreEqual("IndexColumns", dt.Rows[10][0]);
 			Assert.AreEqual("Indexes", dt.Rows[11][0]);
 
-			if (Is50)
+			if (version >= new Version(5, 0))
 			{
 				Assert.AreEqual("Views", dt.Rows[12][0]);
 				Assert.AreEqual("ViewColumns", dt.Rows[13][0]);
@@ -224,10 +224,11 @@ namespace MySql.Data.MySqlClient.Tests
 			Assert.AreEqual("TINYINT", dt.Rows[3]["DATA_TYPE"].ToString().ToUpper());
 		}
 
-		[Category("5.0")]
 		[Test]
 		public void Procedures()
 		{
+			if (version < new Version(5, 0)) return;
+
 			execSQL("DROP PROCEDURE IF EXISTS spTest");
 			execSQL("CREATE PROCEDURE spTest (id int) BEGIN SELECT 1; END");
 
@@ -240,10 +241,11 @@ namespace MySql.Data.MySqlClient.Tests
 			Assert.AreEqual("spTest", dt.Rows[0][3]);
 		}
 
-		[Category("5.0")]
 		[Test]
 		public void Functions()
 		{
+			if (version < new Version(5, 0)) return;
+
 			execSQL("DROP FUNCTION IF EXISTS spFunc");
 			execSQL("CREATE FUNCTION spFunc (id int) RETURNS INT BEGIN RETURN 1; END");
 
@@ -256,10 +258,11 @@ namespace MySql.Data.MySqlClient.Tests
 			Assert.AreEqual("spFunc", dt.Rows[0][3]);
 		}
 
-		[Category("5.0")]
 		[Test]
 		public void ProcedureParameters()
 		{
+			if (version < new Version(5, 0)) return;
+
 			execSQL("DROP PROCEDURE IF EXISTS spTest");
 			execSQL("CREATE PROCEDURE spTest (id int, name varchar(50)) BEGIN SELECT 1; END");
 
@@ -309,10 +312,11 @@ namespace MySql.Data.MySqlClient.Tests
 			Assert.AreEqual("YES", dt.Rows[1]["IS_RESULT"]);
 		}
 
-		[Category("5.0")]
 		[Test]
 		public void Indexes()
 		{
+			if (version < new Version(5, 0)) return;
+
 			execSQL("DROP TABLE IF EXISTS test");
 			execSQL("CREATE TABLE test (id int, PRIMARY KEY(id))");
 			string[] restrictions = new string[4];
@@ -385,10 +389,11 @@ namespace MySql.Data.MySqlClient.Tests
 			Assert.AreEqual(2, dt.Rows[1]["ORDINAL_POSITION"]);
 		}
 
-		[Category("5.0")]
 		[Test]
 		public void Views()
 		{
+			if (version < new Version(5, 0)) return;
+
 			execSQL("DROP VIEW IF EXISTS vw");
 			execSQL("CREATE VIEW vw AS SELECT Now() as theTime");
 
@@ -401,10 +406,11 @@ namespace MySql.Data.MySqlClient.Tests
 			Assert.AreEqual("vw", dt.Rows[0]["TABLE_NAME"]);
 		}
 
-		[Category("5.0")]
 		[Test]
 		public void ViewColumns()
 		{
+			if (version < new Version(5, 0)) return;
+
 			execSQL("DROP VIEW IF EXISTS vw");
 			execSQL("CREATE VIEW vw AS SELECT Now() as theTime");
 
@@ -420,9 +426,10 @@ namespace MySql.Data.MySqlClient.Tests
 		}
 
 		[Test]
-		[Category("5.0")]
 		public void SingleProcedureParameters()
 		{
+			if (version < new Version(5, 0)) return;
+
 			execSQL("DROP PROCEDURE IF EXISTS spTest");
 			execSQL("CREATE PROCEDURE spTest(id int, IN id2 INT(11), " +
 				"INOUT io1 VARCHAR(20), OUT out1 FLOAT) BEGIN END");
@@ -600,10 +607,11 @@ namespace MySql.Data.MySqlClient.Tests
 			Assert.AreEqual("id", row["REFERENCED_COLUMN_NAME"]);
 		}
 
-		[Category("5.0")]
 		[Test]
 		public void Triggers()
 		{
+			if (version < new Version(5, 0)) return;
+
 			try
 			{
 				suExecSQL("DROP TRIGGER trigger1");
